When saving/loading a text string to S3, also cache it in redis.
This lets us avoid hitting S3 for book summaries every time a book is loaded. This is particularly important when loading books for lists (which don't load the book from redis). Especially since the summary isn't even shown in a list. Why pay for it if we don't need it?