Update vagrant provisioner to add self-signed certs for testing HTTPS.
With this check-in the vagrant provisioner will no longer forward ports on the same IP as the host.
Now, it will set up a new private IP for the guest.

This IP was chosen to hopefully be unused on your network, but be sure to check
the Vagrant file node.vm.network line for network compatibilty. Make local changes
as necessary.

This starts to move towards issue #29